Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
databases==0.9.0
fastapi==0.115.2
gevent==24.10.3
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
numpy==2.1.2
pandas==2.2.3
priority==2.0.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-memcached==1.62
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
tzdata==2024.2
websockets==13.1
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.0
